# swagger/openapi.yaml
...
openapi: 3.0.0
servers:
- url: http://localhost:3000/
info:
  description: 'This is a sample E-Commerce app.  This project requires that you build a fully-functioning e-commerce application REST API that allows users to perform various CRUD operations such as registering an account, browsing products for sale, etc.'
  version: 1.0.0
  title: E-Commerce REST API application
  termsOfService: https://www.codecademy.com/about
  contact:
    email: easoftware@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  "/carts":
    post:
      tags:
      - carts
      summary: Create cart
      description: 'If non-exiatant, done by the app when the user selects a product to purchare.'
      operationId: createCart
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Carts"
        '400':
          description: user_id already used, required value missing
        '404':
          description: cart not inserted
        '409':
          description: user_id violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/CartsPost"
    get:
      tags:
      - carts
      summary: Gets all carts
      description: ''
      operationId: getAllCarts
      responses:
        '200':
          description: successful opertion (No cart rows is OK, array is empty)
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Carts"  
  "/carts/{id}":
    get:
      tags:
      - carts
      summary: Get cart by cart id
      description: ''
      operationId: getCartById
      parameters:
      - name: id
        in: path
        description: 'The cart id that needs to be fetched. Use 2 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Carts"
        '404':
          description: Invalid Cart ID supplied, Cart not found
    put:
      tags:
      - carts
      summary: update cart by cart id
      description: 'This can only be done by the logged in user.'
      operationId: updateCartById
      parameters:
      - name: id
        in: path
        description: 'The cart id of the cart that needs to be updated. Use 2 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Carts"
        '400':
          description: user_id already used, required value missing
        '404':
          description: Invalid Cart ID supplied, Cart not found
        '409':
          description: user_id violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/CartsPut"
    delete:
      tags:
      - carts
      summary: Delete cart
      description: This can only be done by the app or Admin.
      operationId: deleteCart
      parameters:
      - name: id
        in: path
        description: 'The cart id that needs to be deleted. In testing, use before() to generate cart to delete.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Deleted cart - cart's id returned as a string
          content:
            text/plain: 
              schema:
                type: string
        '400': 
          description: unknown error
        '404':
          description: Cart not found
        '409':
          description: Cannot delete - constraint error        
  "/carts/{id}/allItems":
    delete:
      tags:
      - carts
      summary: Delete all cart items on a cart
      description: 'Done by the user.'
      operationId: deleteAllCartItems
      parameters:
      - name: id
        in: path
        description: 'The cart id with cart items to be deleted. In testing, use before() to generate cart id to delete.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Deleted cart items - cart's id returned as a string
          content:
            text/plain: 
              schema:
                type: string
        '404':
          description: Could not delete cart items, invalild cart id
  "/carts/{id}/checkout":
    post:
      tags:
      - carts
      summary: Places an order for user from cart
      description: Done by user. Get payment approval. Moves user's cart and cart items into orders and order items.
      operationId: checkout
      parameters:
      - name: id
        in: path
        description: 'The cart id with with cart items, used to create the order. In testing, use before() to generate cart id to delete. For testing, use quantity 5, user_id 2, and create a test cart with test items to use to checkout.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Orders"
        '400':
          description: could not remove cart, wrong number of cart items removed, could not move cart items into order items, could not create new orders from cart, could not get total price from cart
        '404':
          description: Invalid Cart ID supplied, could not place order
      requestBody:
        "$ref": "#/components/requestBodies/Carts"
  "/carts/{id}/items":
    post:
      tags:
      - carts
      summary: Create cart item
      description: 'Done when user adds an item to the cart.'
      operationId: createCartItem
      parameters:
      - name: id
        in: path
        description: 'The cart id link the cart item. Use 1 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Cart_items"
        '400':
          description: product id not valid, required value missing
        '404':
          description: cart not found
        '409':
          description: cart_id or product_id violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/CartItemsPost"
    get:
      tags:
      - carts
      summary: Get cart items by cart id
      description: ''
      operationId: getCartItemsById
      parameters:
      - name: id
        in: path
        description: 'The cart id for items to be fetched. Use 1 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Cart_items"
        '404':
          description: Invalid Cart ID supplied, Cart not found
  "/carts/items/{itemId}":
    get:
      tags:
      - carts
      summary: Get cart item by cart item id
      description: ''
      operationId: getCartItemById
      parameters:
      - name: itemId
        in: path
        description: 'The cart item id that needs to be fetched. Use 2 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Cart_items"
        '404':
          description: Invalid Cart Item ID supplied, cart item not found
    put:
      tags:
      - carts
      summary: Get cart item by cart item id
      description: ''
      operationId: getCartItemByCartId
      parameters:
      - name: itemId
        in: path
        description: 'The cart item id for items to be fetched. Use 2 for testing.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Cart_items"
        '400':
          description: product id not valid, required value missing
        '404':
          description: Invalid Cart Item ID supplied, Cart Item not found
        '409':
          description: product_id violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/CartItemsPut"
    delete:
      tags:
      - carts
      summary: Delete cart item
      description: Done by the user
      operationId: deleteCartItem
      parameters:
      - name: itemId
        in: path
        description: 'The cart item id that needs to be deleted. In testing, use before() to generate cart item to delete.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Deleted cart item - cart item's id returned as a string
          content:
            text/plain: 
              schema:
                type: string
        '404':
          description: Cart item not found, invalid cart item id
...
components:
  schemas:
    Carts:
      type: object
      properties:
        id:
          type: integer
          format: int32
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        used_id:
          type: integer
          format: int32
    Cart_items:
      type: object
      properties:
        id:
          type: integer
          format: int32
        user_id:
          type: integer
          format: int32
        product_id:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
...
    Orders:
      type: object
      properties:
        id:
          type: integer
          format: int32
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        status:
          type: string
        total_price:
          type: number
          format: Decimal
        user_id:
          type: integer
          format: int32
...
  requestBodies:
    CartItemsPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              cart_id:
                type: integer
                format: int32
              product_id:
                type: integer
                format: int32
              quantity:
                type: integer
                format: int32
    CartItemsPut:
      content:
        application/json:
          schema:
            type: object
            properties:
              product_id:
                type: integer
                format: int32
              quantity:
                type: integer
                format: int32
    Carts:
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                type: integer
                format: int32
              created:
                type: string
                format: date-time
              modified:
                type: string
                format: date-time
              used_id:
                type: integer
                format: int32
    CartsPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              created:
                type: string
                format: date-time
              used_id:
                type: integer
                format: int32
    CartsPut:
      content:
        application/json:
          schema:
            type: object
            properties:
              modified:
                type: string
                format: date-time
              used_id:
                type: integer
                format: int32
...